+++ src/components/poster_image/index.js
@@ -16,7 +16,14 @@ import styles from './styles';
 // utils
 import playYoutubeVideo from '../../utils/playYoutubeVideo';
 
-export default function PosterImage({videoUrl, image}) {
+export default function PosterImage({
+  videoUrl,
+  image,
+  item,
+  addToFavorites,
+  deleteToFavorites,
+  isFavorite,
+}) {
   if (videoUrl) {
     return (
       <View style={styles.containerPosterImage}>
@@ -37,12 +44,21 @@ export default function PosterImage({videoUrl, image}) {
             <Image source={PlayButton} style={styles.playImage} />
           </TouchableOpacity>
         </ImageBackground>
-        <TouchableOpacity
-          onPress={() => alert('Hi')}
-          style={styles.addMyFavoritesButton}>
-          <Icon style={styles.addIcon} name="add" />
-          <Text style={styles.addMyFavoritesText}>Add to Favorites</Text>
-        </TouchableOpacity>
+        {isFavorite ? (
+          <TouchableOpacity
+            onPress={() => deleteToFavorites()}
+            style={styles.addMyFavoritesButton}>
+            <Icon style={styles.addIcon} type="Ionicons" name="heart" />
+            <Text style={styles.addMyFavoritesText}>Delete to Favorites</Text>
+          </TouchableOpacity>
+        ) : (
+          <TouchableOpacity
+            onPress={() => addToFavorites()}
+            style={styles.addMyFavoritesButton}>
+            <Icon style={styles.addIcon} name="heart-empty" />
+            <Text style={styles.addMyFavoritesText}>Add to Favorites</Text>
+          </TouchableOpacity>
+        )}
       </View>
     );
   } else {
@@ -53,12 +69,21 @@ export default function PosterImage({videoUrl, image}) {
           imageStyle={styles.img}
           resizeMode="stretch"
           source={image}>
-          <TouchableOpacity
-            onPress={() => alert('Hi')}
-            style={styles.addMyFavoritesButton2}>
-            <Icon style={styles.addIcon} name="add" />
-            <Text style={styles.addMyFavoritesText}>Add to Favorites</Text>
-          </TouchableOpacity>
+          {isFavorite ? (
+            <TouchableOpacity
+              onPress={() => deleteToFavorites()}
+              style={styles.addMyFavoritesButtonDelete}>
+              <Icon style={styles.addIcon} type="Ionicons" name="heart" />
+              <Text style={styles.addMyFavoritesText}>Delete to Favorites</Text>
+            </TouchableOpacity>
+          ) : (
+            <TouchableOpacity
+              onPress={() => addToFavorites()}
+              style={styles.addMyFavoritesButton2}>
+              <Icon style={styles.addIcon} name="heart-empty" />
+              <Text style={styles.addMyFavoritesText}>Add to Favorites</Text>
+            </TouchableOpacity>
+          )}
         </ImageBackground>
       </View>
     );