+++ src/routes/app/index.js
@@ -6,12 +6,14 @@ import {createStackNavigator} from 'react-navigation-stack';
 import Home from '../../screens/app/Home';
 import DetailData from '../../screens/app/DetailData';
 import Search from '../../screens/app/Search';
+import Favorites from '../../screens/app/Favorites';
 
 const appStackNavigator = createStackNavigator(
   {
     Home,
     DetailData,
     Search,
+    Favorites,
   },
   {
     initialRouteName: 'Home',