+++ src/screens/app/Favorites/styles.js
@@ -1,0 +1,15 @@
+import {StyleSheet} from 'react-native';
+import {resize} from '../../../utils/styles';
+import colors from '../../../utils/colors';
+
+export default StyleSheet.create({
+  list: {
+    padding: 10,
+  },
+  colorIcon: {
+    color: colors.primaryLight,
+  },
+  colorCloseIcon: {
+    color: colors.backgroundFinished,
+  },
+});